+++ packages/react-final-form-schema/src/fields/dynamic/component.tsx
@@ -1,9 +1,18 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import {
+  useForm as defaultUseForm,
   useFormState as defaultUseFormState,
 } from 'react-final-form';
+import {
+  useEffect as defaultUseEffect,
+  useRef as defaultUseRef,
+} from 'react';
 import type { ReactElement } from 'react';
+
+import {
+  useFormSchemaState as defaultUseFormSchemaState,
+} from '../../core';
 
 import type {
   FieldComponentProps,
@@ -33,7 +42,11 @@ Payload,
     /**
      * For tests only
      */
+    useForm?: typeof defaultUseForm;
     useFormState?: typeof defaultUseFormState;
+    useFormSchemaState?: typeof defaultUseFormSchemaState;
+    useEffect?: typeof defaultUseEffect;
+    useRef?: typeof defaultUseRef;
   };
 
 export function DynamicField<
@@ -44,12 +57,17 @@ SerializedValues extends Record<string, any>,
 Errors extends Record<string, any>,
 Payload,
 >({
+  name,
   fieldSchema,
 
   getFieldSchema,
   getFieldType,
 
+  useForm = defaultUseForm,
   useFormState = defaultUseFormState,
+  useFormSchemaState = defaultUseFormSchemaState,
+  useEffect = defaultUseEffect,
+  useRef = defaultUseRef,
 
   ...rest
 }: DynamicFieldProps<
@@ -60,12 +78,20 @@ SerializedValues,
 Errors,
 Payload
 >): ReactElement {
+  const form = useForm<Values>();
+
   const {
     values,
   } = useFormState<Values, Values>();
 
   const {
+    isValuesReady,
+  } = useFormSchemaState();
+
+  const {
     getSchema,
+    onShow,
+    onHide,
   } = fieldSchema as unknown as DynamicSchema<
   FieldSchema,
   Values,
@@ -75,6 +101,32 @@ Payload
   >;
 
   const schema = getSchema(values, 'render', getFieldSchema, getFieldType);
+
+  const isFirstRenderRef = useRef(true);
+
+  useEffect(() => {
+    if (isValuesReady) {
+      if (isFirstRenderRef.current) {
+        isFirstRenderRef.current = false;
+        return;
+      }
+
+      if (schema) {
+        if (onShow) {
+          onShow(form, name, schema, getFieldSchema, getFieldType);
+        }
+
+        return;
+      }
+
+      if (onHide) {
+        onHide(form, name, getFieldSchema, getFieldType);
+        return;
+      }
+
+      isFirstRenderRef.current = true;
+    }
+  }, [Boolean(schema)]);
 
   if (!schema) {
     return null;
@@ -89,6 +141,7 @@ Payload
   return (
     <FieldComponent
       {...rest}
+      name={name}
       fieldSchema={schema}
       getFieldSchema={getFieldSchema}
       getFieldType={getFieldType}