+++ packages/react-final-form-schema/src/fields/dynamic/schema.ts
@@ -1,5 +1,9 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import type {
+  FormApi,
+} from 'final-form';
+
+import type {
   GetFieldSchema,
   GetFieldType,
   PhaseType,
@@ -12,10 +16,53 @@ RawValues extends Record<string, any> = Record<string, any>,
 SerializedValues extends Record<string, any> = Record<string, any>,
 Errors extends Record<string, any> = Record<string, any>,
 > = {
-  getSchemaAsync?: (
+  /**
+   * Callback that should return schema of field or `null` if field can't be shown
+   */
+  getSchema: (
+    /**
+     * object of values of form, type depends from `phase` (2nd argument)
+     */
     values: Values | RawValues,
+    /**
+     * current phase
+     */
     phase: PhaseType,
+    /**
+     * current `getFieldSchema`
+     */
     getFieldSchema: GetFieldSchema<FieldSchema>,
+    /**
+     * global `getFieldType`
+     */
+    getFieldType: GetFieldType<
+    FieldSchema,
+    Values,
+    RawValues,
+    SerializedValues,
+    Errors
+    >,
+  ) => FieldSchema;
+
+  /**
+   * Async callback that can be used instead of `getSchema` on parsing phase
+   */
+  getSchemaAsync?: (
+    /**
+     * object of values of form, type depends from `phase` (2nd argument)
+     */
+    values: Values | RawValues,
+    /**
+     * current phase
+     */
+    phase: PhaseType,
+    /**
+     * current `getFieldSchema`
+     */
+    getFieldSchema: GetFieldSchema<FieldSchema>,
+    /**
+     * global `getFieldType`
+     */
     getFieldType: GetFieldType<
     FieldSchema,
     Values,
@@ -25,10 +72,29 @@ Errors extends Record<string, any> = Record<string, any>,
     >,
   ) => Promise<FieldSchema>;
 
-  getSchema: (
-    values: Values | RawValues,
-    phase: PhaseType,
+  /**
+   * callback than called when field has shown
+   */
+  onShow?: (
+    /**
+     * instance of `final-form`
+     */
+    form: FormApi<Values>,
+    /**
+     * name of field
+     */
+    name: string,
+    /**
+     * result schema of subfield
+     */
+    schema: FieldSchema,
+    /**
+     * current `getFieldSchema`
+     */
     getFieldSchema: GetFieldSchema<FieldSchema>,
+    /**
+     * global `getFieldType`
+     */
     getFieldType: GetFieldType<
     FieldSchema,
     Values,
@@ -36,5 +102,33 @@ Errors extends Record<string, any> = Record<string, any>,
     SerializedValues,
     Errors
     >,
-  ) => FieldSchema;
+  ) => void;
+
+  /**
+   * callback than called when field has hidden
+   */
+  onHide?: (
+    /**
+     * instance of `final-form`
+     */
+    form: FormApi<Values>,
+    /**
+     * name of field
+     */
+    name: string,
+    /**
+     * current `getFieldSchema`
+     */
+    getFieldSchema: GetFieldSchema<FieldSchema>,
+    /**
+     * global `getFieldType`
+     */
+    getFieldType: GetFieldType<
+    FieldSchema,
+    Values,
+    RawValues,
+    SerializedValues,
+    Errors
+    >,
+  ) => void;
 };